added output doc
